# Backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_foydalanuvchilar-bazasi_backend_internal_modules_survey_dto.CreateSurveyRequest:
    properties:
      description:
        maxLength: 10000
        type: string
      questions:
        items:
          type: integer
        type: array
      settings:
        items:
          type: integer
        type: array
      slug:
        maxLength: 100
        type: string
      sortOrder:
        type: integer
      title:
        maxLength: 255
        type: string
    required:
    - questions
    - slug
    - title
    type: object
  github_com_foydalanuvchilar-bazasi_backend_internal_modules_survey_dto.ErrorResponse:
    properties:
      error:
        example: validation failed
        type: string
      message:
        example: Xatolik yuz berdi
        type: string
      success:
        example: false
        type: boolean
    type: object
  github_com_foydalanuvchilar-bazasi_backend_internal_modules_survey_dto.FormSubmitRequest:
    properties:
      answers:
        items:
          type: integer
        type: array
      name:
        example: Alisher Karimov
        maxLength: 200
        minLength: 2
        type: string
      phone:
        example: "+998901234567"
        maxLength: 30
        minLength: 9
        type: string
    required:
    - answers
    - name
    - phone
    type: object
  github_com_foydalanuvchilar-bazasi_backend_internal_modules_survey_dto.MessageResponse:
    properties:
      data: {}
      message:
        example: Muvaffaqiyatli
        type: string
      success:
        example: true
        type: boolean
    type: object
  github_com_foydalanuvchilar-bazasi_backend_internal_modules_survey_dto.UpdateSurveyRequest:
    properties:
      description:
        maxLength: 10000
        type: string
      questions:
        items:
          type: integer
        type: array
      settings:
        items:
          type: integer
        type: array
      slug:
        maxLength: 100
        type: string
      sortOrder:
        type: integer
      title:
        maxLength: 255
        type: string
    required:
    - questions
    - slug
    - title
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: Modular monolit backend. Admin, Company, Survey va Settings modullari.
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: Foydalanuvchilar bazasi API
  version: "1.0"
paths:
...
  /company/surveys/{id}/responses:
    get:
      description: Berilgan so'rovnoma (slug yoki id) javoblari
      parameters:
      - description: Survey slug yoki UUID
        in: path
        name: id
        required: true
        type: string
      - default: 1
        description: Sahifa
        in: query
        name: page
        type: integer
      - default: 20
        description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_survey_dto.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_survey_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_survey_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Bitta so'rovnoma javoblari
      tags:
      - Survey Responses
  /company/surveys/{id}/responses/summary:
    get:
      parameters:
      - description: Survey slug yoki UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_survey_dto.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_survey_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: So'rovnoma javoblari statistikasi
      tags:
      - Survey Responses
  /company/surveys/responses:
    get:
      description: Kompaniyaning barcha so'rovnomalariga kelgan javoblar (name, phone
        bilan)
      parameters:
      - default: 1
        description: Sahifa
        in: query
        name: page
        type: integer
      - default: 20
        description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_survey_dto.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_survey_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Barcha so'rovnoma javoblari
      tags:
      - Survey Responses
  /company/surveys/responses/{responseId}:
    delete:
      parameters:
      - description: Response UUID
        in: path
        name: responseId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_survey_dto.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_survey_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Javobni o'chirish
      tags:
      - Survey Responses
    get:
      parameters:
      - description: Response UUID
        in: path
        name: responseId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_survey_dto.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_survey_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Bitta javobni olish
      tags:
      - Survey Responses
  /forms/{slug}:
    get:
      description: Public form frontend uchun. Auth kerak emas. published yoki closed.
      parameters:
      - description: Survey slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_survey_dto.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_survey_dto.ErrorResponse'
      summary: Forma — so'rovnomani olish (tokensiz)
      tags:
      - Form
    post:
      consumes:
      - application/json
      description: name (ism) va phone (telefon) majburiy. Auth kerak emas. Faqat
        published.
      parameters:
      - description: Survey slug
        in: path
        name: slug
        required: true
        type: string
      - description: Ism, telefon va javoblar
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_survey_dto.FormSubmitRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_survey_dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_survey_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_survey_dto.ErrorResponse'
      summary: Forma — so'rovnomani topshirish (tokensiz)
      tags:
      - Form
...
  /surveys:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_survey_dto.MessageResponse'
      summary: Published surveys
      tags:
      - Surveys
    post:
      consumes:
      - application/json
      description: Kompaniya yangi so'rovnoma yaratadi. Slug va public link avtomatik
        hosil bo'ladi.
      parameters:
      - description: So'rovnoma ma'lumotlari
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_survey_dto.CreateSurveyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_survey_dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_survey_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_survey_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_survey_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Yangi so'rovnoma yaratish
      tags:
      - Surveys
  /surveys/{id}:
    delete:
      description: So'rovnomani tizimdan o'chiradi
      parameters:
      - description: Survey UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_survey_dto.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_survey_dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_survey_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_survey_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_survey_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: So'rovnomani o'chirish
      tags:
      - Surveys
    get:
      description: Bitta so'rovnoma ma'lumotlari va public linkini qaytaradi
      parameters:
      - description: Survey UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_survey_dto.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_survey_dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_survey_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_survey_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_survey_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: So'rovnomani ID bo'yicha olish
      tags:
      - Surveys
    put:
      consumes:
      - application/json
      description: Title, description, is_active yangilanadi. Slug va link o'zgarmaydi.
      parameters:
      - description: Survey UUID
        in: path
        name: id
        required: true
        type: string
      - description: Yangilangan ma'lumotlar
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_survey_dto.UpdateSurveyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_survey_dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_survey_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_survey_dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_survey_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_survey_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_foydalanuvchilar-bazasi_backend_internal_modules_survey_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: So'rovnomani yangilash
      tags:
      - Surveys
securityDefinitions:
  BearerAuth:
    description: 'JWT token. Format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"